Задача 2 . /*добавил проверку на расширение файла. копироваться будут только файлы с расширением txt

+++ lesson5/src/main/java/ru/geekbrains/core/lesson5/Program.java
@@ -15,10 +15,10 @@ public class Program {
 
 
     /**
-     1.  Создать 2 текстовых файла, примерно по 50-100 символов в каждом(особого значения не имеет);
-     2.  Написать метод, «склеивающий» эти файлы, то есть вначале идет текст из первого файла, потом текст из второго.
-     3.* Написать метод, который проверяет, присутствует ли указанное пользователем слово в файле (работаем только с латиницей).
-     4.* Написать метод, проверяющий, есть ли указанное слово в папке
+     * 1.  Создать 2 текстовых файла, примерно по 50-100 символов в каждом(особого значения не имеет);
+     * 2.  Написать метод, «склеивающий» эти файлы, то есть вначале идет текст из первого файла, потом текст из второго.
+     * 3.* Написать метод, который проверяет, присутствует ли указанное пользователем слово в файле (работаем только с латиницей).
+     * 4.* Написать метод, проверяющий, есть ли указанное слово в папке
      */
 
     private static final Random random = new Random();
@@ -27,31 +27,28 @@ public class Program {
     private static final String TO_SEARCH = "GeekBrains";
 
 
-
     public static void main(String[] args) throws IOException {
         writeFileContents("sample01.txt", 60, 5);
         writeFileContents("sample02.txt", 90, 5);
         concatenate("sample01.txt", "sample02.txt", "sample01_out.txt");
 
 
-
-
         long i = 0;
-        while ( (i = searchInFile("sample01_out.txt", i, TO_SEARCH)) > 0){
+        while ((i = searchInFile("sample01_out.txt", i, TO_SEARCH)) > 0) {
             System.out.printf("Файл содержит искомое слово, смещение:  %d\n", i);
         }
         System.out.println("Завершение поиска.");
 
 
         String[] fileNames = new String[10];
-        for (int j = 0; j < fileNames.length; j++){
+        for (int j = 0; j < fileNames.length; j++) {
             fileNames[j] = "file_" + j + ".txt";
             writeFileContents(fileNames[j], 30, 3);
             System.out.printf("Файл %s был создан.\n", fileNames[j]);
         }
 
         List<String> result = searchMatch(new File("."), TO_SEARCH);
-        for (String s : result){
+        for (String s : result) {
             System.out.printf("Файл %s содержит искомое слово '%s'\n", s, TO_SEARCH);
         }
         createBackup(new File("."));
@@ -60,25 +57,27 @@ public class Program {
 
     /**
      * Метод генерации некоторой последовательности символов
+     *
      * @param amount кол-во символов
      * @return последовательность символов
      */
-    private static String generateSymbols(int amount){
+    private static String generateSymbols(int amount) {
         StringBuilder stringBuilder = new StringBuilder();
-        for (int i = 0; i < amount; i++){
-            stringBuilder.append((char)random.nextInt(CHAR_BOUND_L, CHAR_BOUND_H + 1));
+        for (int i = 0; i < amount; i++) {
+            stringBuilder.append((char) random.nextInt(CHAR_BOUND_L, CHAR_BOUND_H + 1));
         }
         return stringBuilder.toString();
     }
 
     /**
      * Записать последовательность символов в файл
+     *
      * @param fileName имя файла
-     * @param length длина последовательности символов
+     * @param length   длина последовательности символов
      * @throws IOException
      */
     private static void writeFileContents(String fileName, int length) throws IOException {
-        try(FileOutputStream fileOutputStream = new FileOutputStream(fileName)){
+        try (FileOutputStream fileOutputStream = new FileOutputStream(fileName)) {
             fileOutputStream.write(generateSymbols(length).getBytes(StandardCharsets.UTF_8));
         }
     }
@@ -86,18 +85,18 @@ public class Program {
     /**
      * Записать последовательность символов в файл, при этом, случайным образом дописать осознанное слово
      * для поиска
+     *
      * @param fileName имя файла
-     * @param length длина последовательности символов
-     * @param words кол-во слов для поиска
+     * @param length   длина последовательности символов
+     * @param words    кол-во слов для поиска
      * @throws IOException
      */
     private static void writeFileContents(String fileName, int length, int words) throws IOException {
-        try(FileOutputStream fileOutputStream = new FileOutputStream(fileName)){
-            for(int i = 0; i < words; i++){
-                if (random.nextInt(4) == 0){
+        try (FileOutputStream fileOutputStream = new FileOutputStream(fileName)) {
+            for (int i = 0; i < words; i++) {
+                if (random.nextInt(4) == 0) {
                     fileOutputStream.write(TO_SEARCH.getBytes(StandardCharsets.UTF_8));
-                }
-                else {
+                } else {
                     fileOutputStream.write(generateSymbols(length).getBytes(StandardCharsets.UTF_8));
                 }
             }
@@ -106,44 +105,42 @@ public class Program {
 
     private static void concatenate(String fileIn1, String fileIn2, String fileOut) throws IOException {
         // На запись
-        try(FileOutputStream fileOutputStream = new FileOutputStream(fileOut)){
+        try (FileOutputStream fileOutputStream = new FileOutputStream(fileOut)) {
             int c;
             // На чтение
-            try(FileInputStream fileInputStream = new FileInputStream(fileIn1)) {
-                while ( (c = fileInputStream.read()) != -1){
+            try (FileInputStream fileInputStream = new FileInputStream(fileIn1)) {
+                while ((c = fileInputStream.read()) != -1) {
                     fileOutputStream.write(c);
                 }
             }
             // На чтение
-            try(FileInputStream fileInputStream = new FileInputStream(fileIn2)) {
-                while ( (c = fileInputStream.read()) != -1){
+            try (FileInputStream fileInputStream = new FileInputStream(fileIn2)) {
+                while ((c = fileInputStream.read()) != -1) {
                     fileOutputStream.write(c);
                 }
             }
         }
     }
 
-    private static long searchInFile(String fileName, String search) throws IOException{
+    private static long searchInFile(String fileName, String search) throws IOException {
         return searchInFile(fileName, 0, search);
     }
 
-    private static long searchInFile(String fileName, long offset, String search) throws IOException{
-        try(FileInputStream fileInputStream = new FileInputStream(fileName)){
+    private static long searchInFile(String fileName, long offset, String search) throws IOException {
+        try (FileInputStream fileInputStream = new FileInputStream(fileName)) {
             fileInputStream.skipNBytes(offset);
             byte[] searchData = search.getBytes(StandardCharsets.UTF_8);
             int c;
             int i = 0;
-            while ((c = fileInputStream.read())  != -1){
-                if (c == searchData[i]){
+            while ((c = fileInputStream.read()) != -1) {
+                if (c == searchData[i]) {
                     i++;
-                }
-                else{
+                } else {
                     i = 0;
                     if (c == searchData[i])
                         i++;
                 }
-                if (i == searchData.length)
-                {
+                if (i == searchData.length) {
                     return offset;
                 }
                 offset++;
@@ -157,9 +154,9 @@ public class Program {
         File[] files = file.listFiles();
         if (files == null)
             return list;
-        for (int i = 0; i < files.length; i++){
-            if (files[i].isFile()){
-                if (searchInFile(files[i].getCanonicalPath(), search) > -1){
+        for (int i = 0; i < files.length; i++) {
+            if (files[i].isFile()) {
+                if (searchInFile(files[i].getCanonicalPath(), search) > -1) {
                     list.add(files[i].getCanonicalPath());
                 }
             }
@@ -167,38 +164,43 @@ public class Program {
         return list;
     }
 
-    /**Домашнее задание
+    /**
+     * Домашнее задание
      * Написать функцию, создающую резервную копию всех файлов в директории во вновь созданную папку ./backup
      * 1.создам метод createBackup
      * 2.метод будет создавать директорию Backup , если она не создана
      * 3.получим список всех файлов в родительской  директории
-     * 4.копирую файлы в созданую папку ./backup*/
+     * 4.копирую файлы в созданую папку ./backup
+     */
 
     public static void createBackup(File sourceDir) throws IOException {
-        File backupDir = new File(sourceDir,"backup");
-        if (!backupDir.exists()){
-            if(!backupDir.mkdir()){
+        File backupDir = new File(sourceDir, "backup");
+        if (!backupDir.exists()) {
+            if (!backupDir.mkdir()) {
                 throw new IOException("не удалось создать папку для копирования");
             }
         }
 
-        File[] files =sourceDir.listFiles();
-        if(files==null){
+        File[] files = sourceDir.listFiles();
+        if (files == null) {
             throw new IOException("Каталог пуст");
         }
 
-        for (File file: files){
-            File destFile = new File(backupDir, file.getName());
-            /*StandardCopyOption.REPLACE_EXISTING - заменяет файл, если он уже есть */
-            Files.copy(file.toPath(),destFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
-            System.out.printf("Файл %s\n успешно скопирован",file.getName());
+        for (File file : files) {
+            /*добавил проверку на расширение файла. копироваться будут только файлы с расширением txt
+            * file.isFile()-проверяет явлектся ли обьект файлом ,а не папкой
+            * file.getName() -получаем строку с именем файла
+            * .endsWith(".txt") - проеряем заканчивается ли эта строка на ".txt"
+            * */
+            if (file.isFile() && file.getName().endsWith(".txt")) {
+                File destFile = new File(backupDir, file.getName());
+                /*StandardCopyOption.REPLACE_EXISTING - заменяет файл, если он уже есть */
+                Files.copy(file.toPath(), destFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                System.out.printf("Файл %s\n успешно скопирован ", file.getName());
 
+            }
         }
     }
-
-
-
-
 
 
 }
